Organize functions into appropiate go files
