+++ pkg/onit/cli/run.go
@@ -19,7 +19,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/onosproject/onos-test/pkg/onit/k8s"
+	"github.com/onosproject/onos-test/pkg/onit/setup"
 
 	"github.com/onosproject/onos-test/test"
 
@@ -165,12 +165,6 @@ func getRunBenchSuiteCommand(registry *runner.TestRegistry) *cobra.Command {
 }
 
 func runTestsRemote(cmd *cobra.Command, testID string, commandType string, tests []string, count int) {
-	// Get the onit controller
-
-	controller, err := k8s.NewController()
-	if err != nil {
-		exitError(err)
-	}
 
 	// Get the cluster ID
 	clusterID, err := cmd.Flags().GetString("cluster")
@@ -178,8 +172,10 @@ func runTestsRemote(cmd *cobra.Command, testID string, commandType string, tests
 		exitError(err)
 	}
 
-	// Get the cluster controller
-	cluster, err := controller.GetCluster(clusterID)
+	testSetupBuilder := setup.New()
+	testSetupBuilder.SetClusterID(clusterID)
+	testSetup := testSetupBuilder.Build()
+	cluster, err := testSetup.GetCluster()
 	if err != nil {
 		exitError(err)
 	}
